# insomnia/openapi.yaml
openapi: 3.0.3
info:
  title: Neon Life Drawing API
  description: |
    API for managing life drawing classes, models, venues, and calendar bookings.
    
    ## Authentication
    Most endpoints require a JWT Bearer token. Obtain one via the `/api/auth/login` or `/api/auth/register` endpoints.
    
    ## Base URLs
    - Production: `https://neon-api-production.fliptopbox.workers.dev`
    - Local: `http://localhost:8787`
  version: 1.0.0
  contact:
    name: Neon API Support
servers:
  - url: https://neon-api-production.fliptopbox.workers.dev
    description: Production server
  - url: http://localhost:8787
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or register
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        emailaddress:
          type: string
        active:
          type: integer
        created_on:
          type: string
          format: date-time
        login_on:
          type: string
          format: date-time
        is_admin:
          type: boolean
        fullname:
          type: string
        known_as:
          type: string
        description:
          type: string
    UserBio:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        fullname:
          type: string
        known_as:
          type: string
        description:
          type: string
        instagram:
          type: string
        websites:
          type: array
          items:
            type: string
    Model:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        sex:
          type: integer
          description: "0: Not specified, 1: Male, 2: Female, 3: Non-binary"
        instagram:
          type: string
        portrait:
          type: string
          description: URL to portrait image
        account_holder:
          type: string
        account_number:
          type: string
        account_sortcode:
          type: string
        active:
          type: integer
        created_on:
          type: string
          format: date-time
        modified_on:
          type: string
          format: date-time
        fullname:
          type: string
        known_as:
          type: string
        description:
          type: string
        bio_instagram:
          type: string
        websites:
          type: array
          items:
            type: string
        phone:
          type: string
        email:
          type: string
...
    Image:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        type_id:
          type: integer
        src:
          type: string
          description: Image URL
        active:
          type: integer
        created_on:
          type: string
          format: date-time
        modified_on:
          type: string
          format: date-time
        type_name:
          type: string
...
paths:
...
  /api/users:
    get:
      tags:
        - Users
      summary: List all users (Admin only)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin access required
  /api/users/me:
    get:
      tags:
        - Users
      summary: Get current user profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
  /api/users/me/bio:
    put:
      tags:
        - Users
      summary: Update current user bio
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullname:
                  type: string
                known_as:
                  type: string
                description:
                  type: string
                instagram:
                  type: string
                websites:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Bio updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBio'
        '401':
          description: Unauthorized
  /api/users/{id}/toggle:
    patch:
      tags:
        - Users
      summary: Toggle user active status (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Status toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin access required
  /api/users/{id}/toggle-admin:
    patch:
      tags:
        - Users
      summary: Toggle user admin status (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Admin status toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin access required
  /api/users/{id}:
    delete:
      tags:
        - Users
      summary: Delete user (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin access required

  # Models endpoints
  /api/models:
    get:
      tags:
        - Models
      summary: List all active models
      responses:
        '200':
          description: List of models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Model'
    post:
      tags:
        - Models
      summary: Create or update model profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullname:
                  type: string
                known_as:
                  type: string
                description:
                  type: string
                bio_instagram:
                  type: string
                websites:
                  type: string
                  description: Newline-separated URLs
                sex:
                  type: integer
                instagram:
                  type: string
                portrait:
                  type: string
                account_holder:
                  type: string
                account_number:
                  type: string
                account_sortcode:
                  type: string
      responses:
        '201':
          description: Model created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
        '401':
          description: Unauthorized
  /api/models/by-user/{userId}:
    get:
      tags:
        - Models
      summary: Get model by user ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Model details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
        '404':
          description: Model not found
  /api/models/{id}:
    get:
      tags:
        - Models
      summary: Get model by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Model details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
        '404':
          description: Model not found
    put:
      tags:
        - Models
      summary: Update model
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullname:
                  type: string
                known_as:
                  type: string
                description:
                  type: string
                bio_instagram:
                  type: string
                websites:
                  type: string
                sex:
                  type: integer
                instagram:
                  type: string
                portrait:
                  type: string
                account_holder:
                  type: string
                account_number:
                  type: string
                account_sortcode:
                  type: string
      responses:
        '200':
          description: Model updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
        '401':
          description: Unauthorized
        '404':
          description: Model not found
    delete:
      tags:
        - Models
      summary: Delete model (Admin only, soft delete)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Model deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin access required

  # Artists endpoints
...
  /api/images/user/{userId}:
    get:
      tags:
        - Images
      summary: Get images by user ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of user images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Image'
...